Revert "feat: moving lib to npm package (#12)"
This reverts commit f3fa8f0f5eed0dc8e4a6d286291290e50741bc9c.
