+++ .gitignore
@@ -1,4 +1,2 @@
 # orb.yml is "packed" from source, and not published directly from the repository.
 orb.yml
-dist
-node_modules
