+++ lib/index.js
@@ -1,6 +1,4 @@
-SetupModules() {
-cat <<- "EOF" > "$TMP_DIR"/index.js
-  const { exec } = require('child_process')
+const { exec } = require('child_process')
 const { once } = require('events')
 const MB = 1024 * 1024
 
@@ -102,87 +100,3 @@ global.stopJob = stopJob
 global.$cwd = cwd
 global.shell = shell
 global.$preamble = shellPreamble
-EOF
-cat <<- "EOF" > "$TMP_DIR"/parse_modules.js
-  /**
- * @param {string} modules
- * @returns {{ moduleName: string, alias: string, version: string }[]}
- */
-module.exports = (modules) => {
-  return modules
-    .split('\n')
-    .filter(l => l.trim() !== '')
-    .map(m => {
-        let [module, alias = ''] = m.split('#');
-        let [moduleName = '', version = ''] = module.split(/(?<=.+)@/)
-        alias = alias.trim() === '' ? moduleName : alias.trim()
-        moduleName = moduleName.trim()
-        version = version.trim() === '' ? 'latest' : version.trim()
-        if(moduleName === '') {
-            console.error('No module name provided')
-            return process.exit(1)
-        }
-        return { moduleName, alias, version }
-    })
-}
-EOF
-cat <<- "EOF" > "$TMP_DIR"/install_npm_modules.js
-  const { execSync } = require('child_process')
-const parse_modules = require('./parse_modules')
-
-const modules = parse_modules(process.env.NPM_MODULES)
-    .reduce((acc, { moduleName, version }) => acc += `${moduleName}@${version} `, '')
-
-if (modules === '') process.exit(0)
-execSync(`npm install --no-package-lock --prefix ${process.env.TMP_DIR} ${modules}`)
-EOF
-  node "$TMP_DIR"/install_npm_modules.js
-}
-
-Run() {
-cat <<- "EOF" > "$TMP_DIR"/wrapped.js
-  const parse_modules = require(`${process.env.TMP_DIR}/parse_modules`)
-
-;(async () => {
-  parse_modules(process.env.NPM_MODULES)
-    .map(({ moduleName, alias }) => global[alias] = require(moduleName))
-
-    eval(`
-      (async () => {
-        ${process.env.SCRIPT}
-      })().catch((err) => {
-        console.log(err);
-        process.exit(1);
-      }).then(() => {
-        setTimeout(() => {
-          process.exit(0);
-        }, 1000).unref();
-      })
-    `)
-})().catch((err) => {
-  console.log(err);
-  process.exit(1);
-})
-EOF
-
-  WRAPPED_SCRIPT=$(cat "$TMP_DIR"/wrapped.js)
-
-  # shellcheck disable=SC2016
-  echo "Script to be RUN:"
-  echo "$WRAPPED_SCRIPT"
-  node -r "$TMP_DIR/index.js" -e "$WRAPPED_SCRIPT"
-}
-
-# Will not run if sourced for bats-core tests.
-# View src/tests for more information.
-ORB_TEST_ENV="bats-core"
-if [ "${0#*$ORB_TEST_ENV}" == "$0" ]; then
-  TMP_ROOT=$(dirname "$(mktemp -u)")
-  TMP_DIR="${TMP_ROOT}/ci-node-run-temp"
-  mkdir -p "$TMP_DIR"
-  NODE_PATH="${TMP_DIR}/node_modules"
-  export NODE_PATH
-  export TMP_DIR
-  SetupModules
-  Run
-fi
