--- lib/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "circleci-node-run",
-  "version": "0.0.0-auto",
-  "description": "library to extend circleci functionality within drachovdmytro/node-run orb",
-  "main": "dist/register.js",
-  "scripts": {
-    "build": "tsc",
-    "test": "node --unhandled-rejections=strict ./tests/index.js"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/DrachovDmytro/circleci-node-run.git"
-  },
-  "keywords": [
-    "circleci",
-    "pipeline",
-    "job",
-    "step",
-    "ci"
-  ],
-  "author": "Dmytro Drachov <dmytro.drachov@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/DrachovDmytro/circleci-node-run/issues"
-  },
-  "homepage": "https://github.com/DrachovDmytro/circleci-node-run#readme",
-  "devDependencies": {
-    "@types/node": "^16.0.1",
-    "typescript": "^4.3.5"
-  },
-  "dependencies": {
-    "circleci-typescript-axios": "^0.1.2"
-  }
-}
