+++ lib/test.js
@@ -1,22 +1,15 @@
-const assert = require('assert');
-require('../dist/register')
+const assert = require('assert')
+const index = require('./index.js')
 
 
 ;(async () => await $`exit 1`)()
   .then(_ => assert.fail('Should fail'))
   .catch(err => assert.ok(err.exitCode === 1))
 
+
 ;(async () => await $$`exit 1`)()
   .catch(_ => assert.fail('Should not fail'))
   .then(result => assert.ok(result.exitCode === 1))
-
-;(async () => await $`echo "test"`)()
-  .catch(_ => assert.fail('Should not fail'))
-  .then(({ stdout }) => assert.strictEqual(stdout, 'test'))
-
-;(async () => (await $`echo '{"a": 10}'`).parseJSON())()
-  .catch(_ => assert.fail('Should not fail'))
-  .then((result) => assert.strictEqual(result.a, 10))
 
 
 ;(async () => {
