--- lib/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "outDir": "dist",
-    "module": "CommonJS",
-    "moduleResolution": "Node",
-    "declaration": true,
-  }
-}
