+++ src/@orb.yml
@@ -8,7 +8,7 @@ description: >
 # This information will be displayed in the orb registry and is not mandatory.
 display:
   home_url: "https://www.website.com/docs"
-  source_url: "https://github.com/DrachovDmytro/circleci-node-run#readme"
+  source_url: "https://www.github.com/EXAMPLE_ORG/EXAMPLE_PROJECT"
 
 # If your orb requires other orbs, you can import them like this. Otherwise remove the "orbs" stanza.
 # orbs:
