--- src/commands/file.yml
@@ -1,50 +1,0 @@
-description: >
-  Run node js file
-parameters:
-  file_path:
-    type: string
-    default: ""
-    description: "Path to js file"
-  npm_modules:
-    type: string
-    default: ""
-    description: |
-      npm modules to be installed in format
-        npm_modules: |
-          package_1@semver#alias
-          package_2@semver
-          package_3#alias
-
-        e.g. `lodash#_` module will be available as `_.pick()`
-  step_name:
-    type: string
-    default: "Run NodeJS"
-    description: "Name of step"
-  no_output_timeout:
-    type: string
-    default: "10m"
-    description: "It's circleci's no_output_timeout for script step"
-  background:
-    type: boolean
-    default: false
-    description: "It's circleci's background for script step"
-  working_directory:
-    type: string
-    default: './'
-    description: "It's circleci's working_directory for script step"
-  when:
-    type: enum
-    enum: ["on_success", "on_fail", "always"]
-    default: 'on_success'
-    description: "It's circleci's when for script step"
-steps:
-  - run:
-      no_output_timeout: <<parameters.no_output_timeout>>
-      background: <<parameters.background>>
-      working_directory: <<parameters.working_directory>>
-      when: <<parameters.when>>
-      environment:
-        FILE_PATH: <<parameters.file_path>>
-        NPM_MODULES: <<parameters.npm_modules>>
-      name: <<parameters.step_name>>
-      command: <<include(scripts/file.sh)>>
